Luminara Zoo- ozzoo-simulation/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "luminara-zoo"
version = "0.1.0"
description = "Luminara Zoo — NIT2112 Australian wildlife simulation (OOP + optional Hermes)."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Babatundji Williams-Fulwood" }]

dependencies = []

[project.scripts]
luminara = "main:main"
ozzoo = "main:main"

[project.optional-dependencies]
rich = ["rich>=13,<16"]
hermes = [
  "hermes-agent @ git+https://github.com/NousResearch/hermes-agent.git",
]
ai = [
  "openai>=1.40,<2",
  "google-generativeai>=0.8,<1",
  "python-dotenv>=1.0,<2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["animals", "zoo", "systems", "ui", "hermes_integration"]
